ISSUE #2428: Moved UpcomingMaintenanceWidget to better namespace
